Now differenciates between company and user
falta hacer el componente en sí, pero hago el pr para que quede guardado
