+++ Client/src/Pages/Home/Home_company.jsx
@@ -1,0 +1,3 @@
+export default function HomeCompany() {
+  return <div>company</div>;
+}
